old/requirements.txt
streamlit
pandas
pulp
openpyxl
fuzzywuzzy
XlsxWriter
python-dotenv
pyodbc
requests
urllib.parse